+++ tests/Unit/Colors/Cmyk/ColorTest.php
@@ -102,7 +102,7 @@ final class ColorTest extends BaseTestCase
         $this->assertEquals(
             [1.0, 0.5, 0.2, 0.0, 1],
             array_map(
-                fn(ColorChannelInterface $channel): float => $channel->normalizedValue(),
+                fn(ColorChannelInterface $channel): float => $channel->normalized(),
                 $color->channels(),
             )
         );