+++ tests/Unit/Colors/Hsl/ColorTest.php
@@ -91,7 +91,7 @@ final class ColorTest extends BaseTestCase
         $this->assertEquals(
             [.5, 0.5, 0.25, 1],
             array_map(
-                fn(ColorChannelInterface $channel): float => $channel->normalizedValue(),
+                fn(ColorChannelInterface $channel): float => $channel->normalized(),
                 $color->channels(),
             )
         );