+++ tests/Unit/Colors/Hsv/ColorTest.php
@@ -87,7 +87,7 @@ final class ColorTest extends BaseTestCase
         $this->assertEquals(
             [.5, 0.5, 0.25, 1],
             array_map(
-                fn(ColorChannelInterface $channel): float => $channel->normalizedValue(),
+                fn(ColorChannelInterface $channel): float => $channel->normalized(),
                 $color->channels(),
             )
         );
@@ -96,7 +96,7 @@ final class ColorTest extends BaseTestCase
         $this->assertEquals(
             [.5, 0.5, 0.25, .2],
             array_map(
-                fn(ColorChannelInterface $channel): float => $channel->normalizedValue(),
+                fn(ColorChannelInterface $channel): float => $channel->normalized(),
                 $color->channels(),
             )
         );