+++ tests/Unit/Colors/Oklab/ColorTest.php
@@ -123,7 +123,7 @@ final class ColorTest extends BaseTestCase
         $this->assertEquals(
             [1.0, 0.7500000000000001, 0.0, 1],
             array_map(
-                fn(ColorChannelInterface $channel): float => $channel->normalizedValue(),
+                fn(ColorChannelInterface $channel): float => $channel->normalized(),
                 $color->channels(),
             )
         );
@@ -132,7 +132,7 @@ final class ColorTest extends BaseTestCase
         $this->assertEquals(
             [0, 0.5, 0.5, 1],
             array_map(
-                fn(ColorChannelInterface $channel): float => $channel->normalizedValue(),
+                fn(ColorChannelInterface $channel): float => $channel->normalized(),
                 $color->channels(),
             )
         );
@@ -141,7 +141,7 @@ final class ColorTest extends BaseTestCase
         $this->assertEquals(
             [0, 0.5, 0.5, 0.5019607843137255],
             array_map(
-                fn(ColorChannelInterface $channel): float => $channel->normalizedValue(),
+                fn(ColorChannelInterface $channel): float => $channel->normalized(),
                 $color->channels(),
             )
         );